🛠️ refactor: 코드 리펙토링
- DTO request, response 파일구조 분리
- 가격범위 설정 필터링 요청값 int 로 변경

🔗 Resolves: #130
