+++ backend/JiShop/src/main/java/com/jishop/product/dto/request/ProductRequest.java
@@ -1,4 +1,4 @@
-package com.jishop.product.dto;
+package com.jishop.product.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -19,7 +19,7 @@ public record ProductRequest(
         String sort,
         Long category,
         // 가격 필터 (값 범위: 0~25000, 25000~50000, 50000~100000, 100000~)
-        List<String> priceRanges,
+        List<Integer> priceRanges,
         // 평점 필터 (값 범위: 1~5)
         List<Integer> ratings
 ){
@@ -28,7 +28,7 @@ public record ProductRequest(
                 if (size < 1 || size > 100) {size = 12;}
                 if (sort == null || sort.isEmpty()) {sort = "wish";}
                 if (priceRanges == null || priceRanges.isEmpty()) {
-                        priceRanges = Arrays.asList("0-25000", "25000-50000", "50000-100000", "100000+");
+                        priceRanges = Arrays.asList(0, 25000, 50000, 100000);
                 }
                 if (ratings == null || ratings.isEmpty()) {ratings = Arrays.asList(1, 2, 3, 4, 5);}
         }
