+++ backend/JiShop/src/main/java/com/jishop/product/implementation/ProductQueryHelper.java
@@ -1,7 +1,7 @@
 package com.jishop.product.implementation;
 
 import com.jishop.product.domain.QProduct;
-import com.jishop.product.dto.ProductRequest;
+import com.jishop.product.dto.request.ProductRequest;
 import com.jishop.reviewproduct.domain.QReviewProduct;
 import com.querydsl.core.BooleanBuilder;
