+++ backend/JiShop/src/main/java/com/jishop/product/service/ProductService.java
@@ -1,9 +1,9 @@
 package com.jishop.product.service;
 
 import com.jishop.member.domain.User;
-import com.jishop.product.dto.ProductListResponse;
-import com.jishop.product.dto.ProductRequest;
-import com.jishop.product.dto.ProductResponse;
+import com.jishop.product.dto.response.ProductListResponse;
+import com.jishop.product.dto.request.ProductRequest;
+import com.jishop.product.dto.response.ProductResponse;
 import org.springframework.data.web.PagedModel;
 
 public interface ProductService {
