+++ backend/JiShop/src/main/java/com/jishop/product/service/ProductServiceImpl.java
@@ -5,9 +5,9 @@ import com.jishop.common.exception.ErrorType;
 import com.jishop.member.domain.User;
 import com.jishop.product.domain.Product;
 import com.jishop.product.domain.QProduct;
-import com.jishop.product.dto.ProductListResponse;
-import com.jishop.product.dto.ProductRequest;
-import com.jishop.product.dto.ProductResponse;
+import com.jishop.product.dto.response.ProductListResponse;
+import com.jishop.product.dto.request.ProductRequest;
+import com.jishop.product.dto.response.ProductResponse;
 import com.jishop.product.implementation.ProductQueryHelper;
 import com.jishop.product.repository.ProductRepository;
 import com.jishop.product.repository.ProductRepositoryQueryDsl;
